[REFACTOR] 모달 컴포넌트 코드 리팩토링 및 모듈화
